refactor: migrate test layout to feature-based package structure
Move test classes from layer-based packages (controllers/, services/)
into feature packages (admin/, appointment/, doctor/, patient/, shared/)
to mirror the main source tree. Fix AppointmentService.getAppointments()
to treat the sentinel string "null" as no-filter, matching controller
and frontend conventions (100/100 tests pass).
