+++ app/src/test/java/com/project/back_end/appointment/AppointmentControllerTest.java
@@ -1,4 +1,4 @@
-package com.project.back_end.controllers;
+package com.project.back_end.appointment;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.back_end.appointment.Appointment;
