+++ app/src/test/java/com/project/back_end/appointment/AppointmentServiceTest.java
@@ -1,4 +1,4 @@
-package com.project.back_end.services;
+package com.project.back_end.appointment;
 
 import com.project.back_end.appointment.Appointment;
 import com.project.back_end.appointment.AppointmentDTO;
